Try increasing wait time before hard killing children

+++ lib/Resque/Worker.php
@@ -418,7 +418,6 @@ class Resque_Worker
 	 */
 	public function shutdownNow()
 	{
-		$this->logger->log(Psr\Log\LogLevel::DEBUG, 'SIGINT for {worker}', Array('worker' => $this));
 		$this->shutdown();
 		$this->killChild();
 	}
@@ -435,7 +434,7 @@ class Resque_Worker
 		}
 
 		$this->logger->log(Psr\Log\LogLevel::INFO, 'Killing child at {child}', array('child' => $this->child));
-		/*if(exec('ps -o pid,state -p ' . $this->child, $output, $returnCode) && $returnCode != 1) {
+		if(exec('ps -o pid,state -p ' . $this->child, $output, $returnCode) && $returnCode != 1) {
 			$this->logger->log(Psr\Log\LogLevel::DEBUG, 'Child {child} found, killing.', array('child' => $this->child));
 			posix_kill($this->child, SIGKILL);
 			$this->child = null;
@@ -448,10 +447,6 @@ class Resque_Worker
 		else {
 			$this->logger->log(Psr\Log\LogLevel::INFO, 'Child {child} not found, restarting.', array('child' => $this->child));
 			$this->shutdown();
-		}*/
-
-		if($this->currentJob) {
-			$this->currentJob->recreate();
 		}
 	}
